refactor: consolidate all delete operations into single delete-one tool
Replace separate project, section, and task deletion tools with unified
delete-one tool that accepts type discriminator and entity ID.

Changes:
- Remove: projects-delete-one, sections-delete-one, tasks-delete-one
- Add: delete-one tool with type parameter ('project' | 'section' | 'task')

The consolidated tool provides:
- type='project' + id → Delete project
- type='section' + id → Delete section
- type='task' + id → Delete task

This reduces tool count from 17 to 15 (-2 tools) while maintaining all
functionality. The discriminator approach works well here since deletion
is conceptually the same operation across entity types.

Benefits:
- Consolidates identical interfaces into single, clear tool
- Natural mental model: "delete this type of thing with this ID"
- Maintains all existing functionality
- Continues freeing tool budget for comment functionality
